Added thoughput data aggregation

+++ Gse/src/server/tests/integrity_test.py
@@ -6,6 +6,7 @@ import struct
 import signal
 import random
 import argparse
+import datetime
 
 from threading import Timer, Event
 from subprocess import Popen
@@ -118,6 +119,9 @@ class ServerIntegrityTest:
         time.sleep(passthrough_time_s)    
 
     def monte_carlo_disconnect(self, monte_carlo_time_s):
+        if(monte_carlo_time_s == 0):
+            return
+
         action_limit_s = 60 # Max time before a connect or kill is performed
 
         run_test = Event()
@@ -248,4 +252,47 @@ if __name__ == "__main__":
     test.passthrough(args.pass_time)
 
     test.teardown_class()
+
+
+    # Check logs
+    datetime_str = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S,%f')
+    cmd = "python continuity_log_check.py {}".format(datetime_str)
+    p   = Popen(args=cmd, shell=True)
+    p.wait()
+
+
+    time.sleep(2)
+    # If verbose, aggregate throughput information
+    if(args.verbose):
+        analysis_path = SERVER_CONFIG.get('filepaths','throughput_analysis_filepath')
+
+        final_report_path = os.path.join(analysis_path, "aggregate.txt")
+        final_report = open(final_report_path, "w")
+        final_report.write("Number flight {}\n".format(args.num_flight))
+        final_report.write("Number ground {}\n".format(args.num_ground))
+        final_report.write("Flight throughput {} msg/s\n".format(args.flight_throughput))
+        final_report.write("Flight msg size {} bytes\n".format(args.flight_size))
+        final_report.write("Ground throughput {} msg/s\n".format(args.ground_throughput))
+        final_report.write("Ground msg size {} bytes\n".format(args.ground_size))
+        final_report.write("Montecarlo connect/disconnect time {} seconds\n".format(args.monte_time))
+        final_report.write("Passthrough time {} seconds\n".format(args.pass_time))
+        final_report.write("\n")
+
+        for (path, dirs, files) in os.walk(analysis_path):
+            for dir_name in dirs:
+                dir_path = os.path.join(analysis_path, dir_name)
+                report_path = os.path.join(dir_path, "report.txt")
+                with open(report_path, "r") as f:
+                    for line in f:
+                        final_report.write(line)
+                    final_report.write("\n")
+
+        final_report.close()
+
+
+
+
+
+
+
